Updated Motor Record to release R6-3.
git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/epics/trunk/modules/motor@318 ececa60f-9f63-0410-97c7-ea05888eb260

+++ motorApp/AcsTech80Src/devSPiiPlus.cc
@@ -109,7 +109,7 @@ static struct board_stat **SPiiPlus_cards;
 STATIC long SPiiPlus_init(void *arg)
 {
     long rtnval;
-    long after = (long) arg;
+    int after = (int) arg;
 
     if (after == 0)
     {
