+++ motorApp/MclennanSrc/devPM304.cc
@@ -115,7 +115,7 @@ static struct board_stat **PM304_cards;
 STATIC long PM304_init(void *arg)
 {
     long rtnval;
-    long after = (long) arg;
+    int after = (int) arg;
 
     if (after == 0)
     {
